requirement.txt
google-generativeai
elevenlabs
pyaudio
websockets
aiohttp
python-dotenv
sounddevice
